+++ api/src/main/java/org/pmiops/workbench/cohortbuilder/chart/ChartServiceImpl.java
@@ -2,13 +2,11 @@ package org.pmiops.workbench.cohortbuilder.chart;
 
 import com.google.cloud.bigquery.TableResult;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import org.pmiops.workbench.api.BigQueryService;
 import org.pmiops.workbench.cohortbuilder.ParticipantCriteria;
 import org.pmiops.workbench.cohortbuilder.mapper.CohortBuilderMapper;
 import org.pmiops.workbench.cohortreview.mapper.CohortReviewMapper;
-import org.pmiops.workbench.exceptions.BadRequestException;
 import org.pmiops.workbench.model.AgeType;
 import org.pmiops.workbench.model.CohortChartData;
 import org.pmiops.workbench.model.DemoChartInfo;
@@ -21,8 +19,6 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class ChartServiceImpl implements ChartService {
-  private static final String BAD_REQUEST_MESSAGE =
-      "Bad Request: Please provide a valid %s. %s is not valid.";
   private final BigQueryService bigQueryService;
 
   private final ChartQueryBuilder chartQueryBuilder;
@@ -65,12 +61,11 @@ public class ChartServiceImpl implements ChartService {
 
   @Override
   public List<DemoChartInfo> findDemoChartInfo(
-      String genderOrSex, String age, SearchRequest request) {
+      GenderOrSexType genderOrSexType, AgeType ageType, SearchRequest request) {
     TableResult result =
         bigQueryService.filterBigQueryConfigAndExecuteQuery(
             chartQueryBuilder.buildDemoChartInfoCounterQuery(
-                new ParticipantCriteria(
-                    request, validateGenderOrSexType(genderOrSex), validateAgeType(age))));
+                new ParticipantCriteria(request, genderOrSexType, ageType)));
 
     return cohortBuilderMapper.tableResultToDemoChartInfo(result);
   }
@@ -101,24 +96,5 @@ public class ChartServiceImpl implements ChartService {
             chartQueryBuilder.buildChartDataQuery(participantId, domain, limit));
 
     return cohortReviewMapper.tableResultToParticipantChartData(result);
-  }
-
-  protected AgeType validateAgeType(String age) {
-    return Optional.ofNullable(age)
-        .map(AgeType::fromValue)
-        .orElseThrow(
-            () ->
-                new BadRequestException(
-                    String.format(BAD_REQUEST_MESSAGE, "age type parameter", age)));
-  }
-
-  protected GenderOrSexType validateGenderOrSexType(String genderOrSex) {
-    return Optional.ofNullable(genderOrSex)
-        .map(GenderOrSexType::fromValue)
-        .orElseThrow(
-            () ->
-                new BadRequestException(
-                    String.format(
-                        BAD_REQUEST_MESSAGE, "gender or sex at birth parameter", genderOrSex)));
   }
 }
